==> Code/5-concat.py <==
"""


parcomp xropen xrcomp  democ autoc polity polity2


"""

# These variables exist in NewWmeasure-full, but not in V-Dem
bdm_extras = [
    "lifeExp", "popWB", "safewater", "hygiene", 
        "measles", "electric", "secondaryschool", 
    "parcomp", "xropen", "xrcomp", "democ", "autoc", 
        "polity", "polity2",
    "dd_Przeworski",
    "gwf_demo", "gwf_demo", "gwf_party", "gwf_military", "gwf_monarchy",
        "gwf_personal", "gwf_nonautocracy",
    "food",
    "transparencyindex",
    "pressRestrict",
    "cpiscore",
    "pop"
]


# These
